[#433] Split extensions
