+++ client/src/commonMain/kotlin/com/github/mustafaozhan/ccc/client/util/CoroutineUtil.kt
@@ -1,0 +1,10 @@
+package com.github.mustafaozhan.ccc.client.util
+
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
+
+fun CoroutineScope.launchIgnored(function: suspend () -> Unit) {
+    launch {
+        function()
+    }
+}
